+++ services/admin-service/package.json
@@ -3,7 +3,6 @@
   "version": "1.0.0",
   "description": "",
   "exports": "./index.js",
-  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node index.js"
