+++ services/manipulation-service/src/image.js
@@ -1,7 +1,6 @@
 const Jimp = require('jimp')
 const faceapi = require('face-api.js')
 const canvas = require('canvas')
-const mysql = require('mysql')
 require('@tensorflow/tfjs-node');
 
 const { Canvas, Image, ImageData } = canvas  
@@ -48,27 +47,5 @@ const overlayHat = async (hat, result, baby) => {
   return await image.getBase64Async(Jimp.MIME_PNG)
 }
 
-const uploadPreview = async (description, hat, preview1, preview2) => {
-  const con = mysql.createConnection({
-    host: process.env.HOST,
-    port: '3306',
-    user: "admin",
-    password: process.env.PASSWORD,
-  });
-
-  con.connect(function(err) {
-      console.log(err)
-      con.query(`INSERT INTO main.images (description, hat, preview1, preview2) VALUES ('${description}', '${hat}', '${preview1}', '${preview2}')`, function(err, result, fields) {
-          if (err) console.log(err);
-          if (result) console.log({description: `${description}`});
-          if (fields) console.log(fields);
-      });
-      // connect to mysql and push data
-  });
-
-  return {description: `${description}`}
-}
-
 exports.findFace = findFace
 exports.overlayHat = overlayHat
-exports.uploadPreview = uploadPreview
